fix(security): systemic Cross-Agency IDOR and Unauthenticated Admin Registration fix(admin): allow agencies to update their users fix(admin): allow update password of passwordless users fix(admin): update password button triggers submit from agency form

+++ backend/src/controllers/userController.ts
@@ -158,6 +158,16 @@ export const create = async (req: Request, res: Response) => {
   const { body }: { body: movininTypes.CreateUserPayload } = req
 
   try {
+    // begin of security check
+    const sessionUserId = req.user?._id
+    const sessionUser = await User.findById(sessionUserId)
+    if (!sessionUser || sessionUser.type == movininTypes.UserType.User) {
+      logger.error(`[user.create] Unauthorized attempt to create user by user ${sessionUserId}`)
+      res.status(403).send('Forbidden: You cannot create user')
+      return
+    }
+    // end of security check
+
     body.verified = false
     body.blacklisted = false
 
@@ -933,6 +943,16 @@ export const update = async (req: Request, res: Response) => {
       return
     }
 
+    // begin of security check
+    const sessionUserId = req.user?._id
+    const sessionUser = await User.findById(sessionUserId)
+    if (!sessionUser || sessionUser.type == movininTypes.UserType.User || (sessionUser.type == movininTypes.UserType.Agency && sessionUserId !== user.agency?.toString())) {
+      logger.error(`[user.update] Unauthorized attempt to update user ${_id} by user ${sessionUserId}`)
+      res.status(403).send('Forbidden: You cannot update user information')
+      return
+    }
+    // end of security check
+
     const {
       fullName,
       phone,
@@ -1260,6 +1280,18 @@ export const changePassword = async (req: Request, res: Response) => {
       return
     }
 
+    // begin of security check
+    const sessionUserId = req.user?._id
+    const sessionUser = await User.findById(sessionUserId)
+    if (!sessionUser 
+      || (sessionUser.type == movininTypes.UserType.User && sessionUserId !== user._id.toString())
+      || (sessionUser.type == movininTypes.UserType.Agency && sessionUserId !== user.agency?.toString())) {
+      logger.error(`[user.changePassword] Unauthorized attempt to update user ${_id} by user ${sessionUserId}`)
+      res.status(403).send('Forbidden: You cannot change user password')
+      return
+    }
+    // end of security check
+
     if (strict && !user.password) {
       logger.error('[user.changePassword] User.password not found:', _id)
       res.sendStatus(204)
@@ -1396,6 +1428,7 @@ export const getUsers = async (req: Request, res: Response) => {
             blacklisted: 1,
             birthDate: 1,
             customerId: 1,
+            agency: 1,
           },
         },
         {
@@ -1433,10 +1466,22 @@ export const deleteUsers = async (req: Request, res: Response) => {
     const { body }: { body: string[] } = req
     const ids: mongoose.Types.ObjectId[] = body.map((id: string) => new mongoose.Types.ObjectId(id))
 
+    const sessionUserId = req.user?._id
+    const sessionUser = await User.findById(sessionUserId)
+
+    let unauthorizedAttemptLogged = false
     for (const id of ids) {
       const user = await User.findById(id)
 
       if (user) {
+        // begin of security check
+        if (!sessionUser || sessionUser.type == movininTypes.UserType.User || (sessionUser.type == movininTypes.UserType.Agency && sessionUserId !== user.agency?.toString())) {
+          logger.error(`[user.delete] Unauthorized attempt to delete user ${id} by user ${sessionUserId}`)
+          unauthorizedAttemptLogged = true
+          continue
+        }
+        // end of security check
+
         await User.deleteOne({ _id: id })
 
         if (user.avatar) {
@@ -1477,6 +1522,11 @@ export const deleteUsers = async (req: Request, res: Response) => {
       } else {
         logger.error('User not found:', id)
       }
+    }
+
+    if (unauthorizedAttemptLogged) {
+      res.status(403).send('Forbidden: You cannot delete some of the users')
+      return
     }
 
     res.sendStatus(200)
